+++ controller/konnect/conditions/conditions.go
@@ -94,3 +94,17 @@ const (
 	// condition type indicating that the KongConsumer reference is invalid.
 	KongConsumerRefReasonInvalid = "Invalid"
 )
+
+const (
+	// KongConsumerGroupRefsValidConditionType is the type of the condition that indicates
+	// whether the KongConsumerGroups referenced by the entity are valid and all point to
+	// existing KongConsumerGroups.
+	KongConsumerGroupRefsValidConditionType = "KongConsumerGroupRefsValid"
+
+	// KongConsumerGroupRefsReasonValid is the reason used with the KongConsumerGroupRefsValid
+	// condition type indicating that all KongConsumerGroup references are valid.
+	KongConsumerGroupRefsReasonValid = "Valid"
+	// KongConsumerGroupRefsReasonInvalid is the reason used with the KongConsumerGroupRefsValid
+	// condition type indicating that one or more KongConsumerGroup references are invalid.
+	KongConsumerGroupRefsReasonInvalid = "Invalid"
+)